requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
numpy==1.20.2
pandas==1.2.3
plotly==4.14.3
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
